requirements.txt
pyspark
kafka-python
hdfs